+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/splunk/DisabledSplunkForwarder.kt
@@ -12,6 +12,10 @@ class DisabledSplunkForwarder : SplunkForwarder {
         return
     }
 
+    override fun jsonifyLog4j(sshConnection: SshConnection, log4jPropertiesPath: String, log4j2ConfigPath: String) {
+        return
+    }
+
     override fun getRequiredPorts(): List<Int> {
         return emptyList()
     }