+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/splunk/SplunkForwarder.kt
@@ -5,5 +5,6 @@ import com.atlassian.performance.tools.ssh.api.SshConnection
 interface SplunkForwarder {
     fun run(sshConnection: SshConnection, name: String, logsPath: String)
     fun jsonifyLog4j(sshConnection: SshConnection, log4jPropertiesPath: String)
+    fun jsonifyLog4j(sshConnection: SshConnection, log4jPropertiesPath: String, log4j2ConfigPath: String)
     fun getRequiredPorts(): List<Int>
 }